Generalized reactable resources
