+++ app/Interfaces/Repositories/ArticleRepositoryInterface.php
@@ -2,11 +2,7 @@
 
 namespace App\Interfaces\Repositories;
 
-use Cog\Contracts\Love\Reactable\Models\Reactable;
-use Cog\Contracts\Love\Reacterable\Models\Reacterable;
-
-interface ArticleRepositoryInterface extends ResourceRepositoryInterface
+interface ArticleRepositoryInterface extends ReactableResourceRepositoryInterface
 {
     public function store($request, $user = null);
-    public function react(Reacterable $user, Reactable $item);
 }
